handle event time as guint32. That's the type we deal with here and it
2005-01-03  Sven Neumann  <sven@gimp.org>

	* app/paint/gimpink.[ch]: handle event time as guint32. That's the
	type we deal with here and it avoids a crash that occured when
	autoscrolling with the Ink tool.

	* app/display/gimpdisplayshell-autoscroll.c: cosmetics.

+++ app/paint/gimpink.h
@@ -54,7 +54,7 @@ struct _GimpInk
   guint32        ts_buffer[TIME_SMOOTHER_BUFFER];
   gint           ts_index;
 
-  gdouble        last_time;     /*  previous time of a motion event      */
+  guint32        last_time;     /*  previous time of a motion event      */
   gdouble        lastx, lasty;  /*  previous position of a motion event  */
 
   gboolean       init_velocity;
